Bug Fixes
boost/parameter/aux_/maybe.hpp: added check for MinGW to use boost::aligned_storage instead of std::aligned_storage.
test/evaluate_category.cpp: split into itself, test/evaluate_category_10.cpp, and test/evaluate_category_16.cpp.
test/macros.cpp: split into itself and macros_eval_category.cpp.
test/preprocessor_eval_category.cpp: split into itself and test/preprocessor_eval_cat_8.cpp
