+++ doc/html/reference.html
@@ -2405,8 +2405,8 @@ href="#boost-parameter-function-call-op-result-tag-namespace-arguments"
 except that the overloaded function call operators and their helper methods
 are <tt class="docutils literal">const</tt>-qualified.</p>
 <p>The <a class="reference external"
-href="../../test/preprocessor_eval_category.cpp"
->test/preprocessor_eval_category.cpp</a> test program demonstrates proper
+href="../../test/preprocessor_eval_cat_8.cpp"
+>test/preprocessor_eval_cat_8.cpp</a> test program demonstrates proper
 usage of this macro.</p>
 </div>
 <div class="section"
@@ -3084,7 +3084,9 @@ r name(
 </dd>
 </dl>
 <p>The <a class="reference external" href="../../test/macros.cpp"
->test/macros.cpp</a> test program demonstrates proper usage of this macro.</p>
+>test/macros.cpp</a> and <a class="reference external"
+href="../../test/macros_eval_category.cpp">test/macros_eval_category.cpp</a>
+test programs demonstrate proper usage of this macro.</p>
 </div>
 <div class="section" id="boost-parameter-keyword-n-k">
 <h2><a class="toc-backref" href="#id61">6.13&nbsp;&nbsp;&nbsp;<tt
