+++ include/boost/parameter/aux_/maybe.hpp
@@ -26,7 +26,8 @@ namespace boost { namespace parameter { namespace aux {
     };
 }}} // namespace boost::parameter::aux
 
-#if defined(BOOST_NO_CXX11_HDR_TYPE_TRAITS) || defined(BOOST_MSVC)
+#if defined(BOOST_NO_CXX11_HDR_TYPE_TRAITS) || defined(BOOST_MSVC) || \
+    defined(__MINGW32__)
 #include <boost/type_traits/aligned_storage.hpp>
 #else
 #include <type_traits>
@@ -38,8 +39,9 @@ namespace boost { namespace parameter { namespace aux {
     // If T is not a reference type, returns a POD which can store T.
     template <class T>
     struct referent_storage
-#if defined(BOOST_NO_CXX11_HDR_TYPE_TRAITS) || defined(BOOST_MSVC)
-        // Some versions of MSVC report that the address() function
+#if defined(BOOST_NO_CXX11_HDR_TYPE_TRAITS) || defined(BOOST_MSVC) || \
+    defined(__MINGW32__)
+        // Some versions of MSVC and MinGW report that the address() function
         // is not a member of std::aligned_storage. -- Cromwell D. Enage
       : boost::aligned_storage<
 #else
